done hw 5

+++ controllers/contacts/removeContact.js
@@ -3,7 +3,8 @@ const { NotFound } = require("http-errors");
 
 const removeContact = async (req, res, next) => {
   const { contactId } = req.params;
-  const result = await Contact.findByIdAndRemove(contactId);
+  const { _id: owner } = req.user;
+  const result = await Contact.findByIdAndRemove({ owner, _id: contactId });
 
   if (!result) {
     throw new NotFound("Not found");
